requirements.txt
crewai
crewai-tools
google-genai
streamlit
python-dotenv
pymongo
PyPDF2
